requirements.txt
python-dotenv=0.18.0
discord.py=1.7.3